changed type of owners variable, byte cause crash when over 255.

+++ UltimateTeam.Toolkit/Models/ItemData.cs
@@ -45,7 +45,7 @@ namespace UltimateTeam.Toolkit.Models
 
         public byte Morale { get; set; }
 
-        public byte Owners { get; set; }
+        public int Owners { get; set; }
 
         public ChemistryStyle PlayStyle { get; set; }
